Add client statistics.
Stats are collected on the client application, when system property
-Dcom.oracle.nosql.sdk.nosqldriver.stats.profile=[regular|full] is used,
for an interval of time, set by
-Dcom.oracle.nosql.sdk.nosqldriver.stats.interval=600 in seconds.

The following stats are collected for each type of request: count,
wire latencies, request and response sizes, errors, retries, retry delay,
rate limit delay. Also if rateLimiting is enabled and concurent connections
number.

StatsConfig also provides an API to programaticaly configure the statistics
collection, to selectively start and stop collection and get access to
collected statistics.