Added logic to manage versions of proxy, kvclient, and kvserver.
Added logic to manage versions of proxy, kvclient, and kvserver based on the
proxy that the driver is connected to. Used for testing, to skip tests that may not
run properly against older versions. No changes to public APIs.

The logic maintains three versions: proxy, kvclient and kvserver. All versions
are first read from the reponse headers and if not there, an environment setting
is used.

Version checks in various tests are done with the minimum of client and server kv versions.
The proxy version is not checked. All three values are available to tests if needed.