Added logic to manage versions of proxy, kvclient, and kvserver. (#14)
Added logic to manage versions of proxy, kvclient, and kvserver based on
envrionment or properties settings. Used for testing, to skip tests that may not
run properly against older versions. No changes to public APIs.

Version checks in various tests are done with the minimum of client and server kv versions.
The proxy version is not checked. All three values are available to tests if needed.