Allow retrying a timed-out query if it does not do writes
The current query driver logic disables retrying a query request if it gets a non-retryable error. But a TimeoutException (or RequestTimeoutException) should allow a retry if the query does not do writes. This change fixes this specific case. Retries will be allowed only if the query already is prepared and the prepare statement does not do writes.