Added backward compatibility with older proxies.
This includes checking for "Invalid query version" from old
servers, and the new UNSUPPORTED_QUERY_VERSION message from new servers.

The client SDK will now maintain a current query version number, similar
to the protocol version number, and if it gets an UnsupportedQueryVersionException
it will attempt to decrement the query version and try again. This follows the same
logic implemented for the overall protocol version.