Add ability to perform a query over distinct subsets of data in parallel
Adds the ability to run a query in parallel if a server supports the feature. For queries that can be run in parallel (not all of them) there is new API that allows an application to specify that a given query request operate over a subset of the data read by the query. This allows multiple coordinated threads, processes, or machines to split up a query to get better performance if there is sufficient throughput allocated